Initial commit for Windmill.
I put the pyselenium stuff we did in the contrib directory in case anyone still using Selenium is interested.

I wrote a new Proxy from the ground up using wsgiref. It currently works as a functioning proxy, although it's slow as dirt because it doesn't do any threading just yet.

-Mikeal



git-svn-id: http://svn.getwindmill.com/trunk@3 78c7df6f-8922-0410-bcd3-9426b1ad491b