We now have a funcitoning wsgi http proxy.
All requests are treated like a normal httpproxy except requests to domain/windmill-serv/, which for now return all files in the core directory.

The proxy doesn't support hop-by-hop headers per the WSGI spec. I've tested against a lot of sites and it works in 99% of all cases.

The server is also threaded, which does a lot to speed it up.

-Mikeal



git-svn-id: http://svn.getwindmill.com/trunk@5 78c7df6f-8922-0410-bcd3-9426b1ad491b