Got the jsonrpc dispatcher working. Got the jsonrpc client working.
Someday we may want to break off the dispatcher and the client into their own module since they are the only ones that have been developed in python so far.

Added option 'usecode' to windmill-admin to force the use of the code module in the interactive shell.

Lots of extra debugging information in the jsonrpc module now.

'shell' now loads xmlrpc and jsonrpc clients before starting for convienience.

-Mikeal




git-svn-id: http://svn.getwindmill.com/trunk@49 78c7df6f-8922-0410-bcd3-9426b1ad491b