Fixes for Leopard.
networksetup is now in /usr/sbin and not hidden in ARDAgent.app.

Leopard broke networksetup -listnetworkserviceorder, I filed a bug with apple. We now use some different logic to get the interface name.

Also, we now get the interface name for the interface that will route to the test url domain instead of the default inteface for all traffic.

-Mikeal



git-svn-id: http://svn.getwindmill.com/trunk@526 78c7df6f-8922-0410-bcd3-9426b1ad491b